+++ packages/db/drizzle/0035_open_the_professor.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `workspace` ADD `limits` text DEFAULT '{}' NOT NULL;
