+++ packages/db/drizzle/meta/_journal.json
@@ -246,6 +246,13 @@
       "when": 1720727898360,
       "tag": "0034_serious_shard",
       "breakpoints": true
+    },
+    {
+      "idx": 35,
+      "version": "6",
+      "when": 1721159796428,
+      "tag": "0035_open_the_professor",
+      "breakpoints": true
     }
   ]
 }
