+++ packages/db/src/schema/index.ts
@@ -2,9 +2,9 @@ export * from "./status_reports";
 export * from "./integration";
 export * from "./pages";
 export * from "./monitors";
+export * from "./workspaces";
 export * from "./users";
 export * from "./page_subscribers";
-export * from "./workspaces";
 export * from "./shared";
 export * from "./notifications";
 export * from "./monitor_status";
