+++ packages/db/src/schema/monitors/constants.ts
@@ -1,52 +1,4 @@
-export const flyRegions = [
-  "ams",
-  "arn",
-  "atl",
-  "bog",
-  "bom",
-  "bos",
-  "cdg",
-  "den",
-  "dfw",
-  "ewr",
-  "eze",
-  "fra",
-  "gdl",
-  "gig",
-  "gru",
-  "hkg",
-  "iad",
-  "jnb",
-  "lax",
-  "lhr",
-  "mad",
-  "mia",
-  "nrt",
-  "ord",
-  "otp",
-  "phx",
-  "qro",
-  "scl",
-  "sjc",
-  "sea",
-  "sin",
-  "syd",
-  "waw",
-  "yul",
-  "yyz",
-] as const;
-
-export const monitorPeriodicity = [
-  "30s",
-  "1m",
-  "5m",
-  "10m",
-  "30m",
-  "1h",
-  "other",
-] as const;
 export const monitorMethods = ["GET", "POST", "HEAD"] as const;
 export const monitorStatus = ["active", "error", "degraded"] as const;
-export const monitorRegions = [...flyRegions] as const;
 
 export const monitorJobTypes = ["website", "cron", "other"] as const;
